--- tests/ui/inline-const/interior-mutable-borrow.stderr
@@ -1,33 +1,0 @@
-error[E0492]: interior mutable shared borrows of temporaries that have their lifetime extended until the end of the program are not allowed
-  --> $DIR/interior-mutable-borrow.rs:18:33
-   |
-LL |     let _: &'static _ = const { &Mutable::new(0u32) };
-   |                                 ^^^^^^^^^^^^^^^^^^^ this borrow of an interior mutable value refers to such a temporary
-   |
-   = note: temporaries in constants and statics can have their lifetime extended until the end of the program
-   = note: to avoid accidentally creating global mutable state, such temporaries must be immutable
-   = help: if you really want global mutable state, try replacing the temporary by an interior mutable `static` or a `static mut`
-
-error[E0492]: interior mutable shared borrows of temporaries that have their lifetime extended until the end of the program are not allowed
-  --> $DIR/interior-mutable-borrow.rs:15:34
-   |
-LL |     let _: &'static _ = &const { Mutable::new(0u32) };
-   |                                  ^^^^^^^^^^^^^^^^^^ this borrow of an interior mutable value refers to such a temporary
-   |
-   = note: temporaries in constants and statics can have their lifetime extended until the end of the program
-   = note: to avoid accidentally creating global mutable state, such temporaries must be immutable
-   = help: if you really want global mutable state, try replacing the temporary by an interior mutable `static` or a `static mut`
-
-error[E0492]: interior mutable shared borrows of temporaries that have their lifetime extended until the end of the program are not allowed
-  --> $DIR/interior-mutable-borrow.rs:21:34
-   |
-LL |     let _: &'static _ = &const { AtomicUsize::new(0) };
-   |                                  ^^^^^^^^^^^^^^^^^^^ this borrow of an interior mutable value refers to such a temporary
-   |
-   = note: temporaries in constants and statics can have their lifetime extended until the end of the program
-   = note: to avoid accidentally creating global mutable state, such temporaries must be immutable
-   = help: if you really want global mutable state, try replacing the temporary by an interior mutable `static` or a `static mut`
-
-error: aborting due to 3 previous errors
-
-For more information about this error, try `rustc --explain E0492`.
